# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Descripción de este servidor backend",
        "title": "API Backend de la aplicación besigabi",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
        "/api/branch/getAll": {
            "get": {
                "description": "Obtiene la lista de todas las sedes estructuradas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener todas las sedes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.SedesNTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/log": {
            "post": {
                "description": "Actualiza el estado log de una taquilla",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Actualizar estado de la taquilla",
                "parameters": [
                    {
                        "description": "Datos del estado de taquilla",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillasEstadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillasEstadoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/logAtencion": {
            "post": {
                "description": "Actualiza el estado de la atención de un turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Actualizar estado de atención",
                "parameters": [
                    {
                        "description": "Datos del estado de atención",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.AtencionEstadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.AtencionEstadoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/pointsOfCare": {
            "get": {
                "description": "Obtiene todos los puntos de atención disponibles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener puntos de atención",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillaNTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/reasons/break": {
            "get": {
                "description": "Obtiene la lista de motivos de descanso (break) para las taquillas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener motivos de descanso",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillaMotivoDescansoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/reasons/cancel": {
            "get": {
                "description": "Obtiene la lista de motivos de cancelación",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener motivos de cancelación",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/reasons/transfer/{idSede}/{idServicio}/{idTurno}": {
            "get": {
                "description": "Obtiene servicios disponibles para la transferencia de un turno",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener servicios para transferencia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Sede",
                        "name": "idSede",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del Servicio",
                        "name": "idServicio",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del Turno",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.ResponseTransfer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/service/getByBranch/{idSede}": {
            "get": {
                "description": "Obtiene los servicios disponibles para una sede específica",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener servicios por sede",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Sede",
                        "name": "idSede",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.ServiciosNTXSedeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/taquillasxSedeServicio/{sede}/{servicio}": {
            "get": {
                "description": "Obtiene las taquillas filtradas por sede y servicio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Obtener taquillas por sede y servicio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Sede",
                        "name": "sede",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del Servicio",
                        "name": "servicio",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillaxSedexServicioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/transfers": {
            "post": {
                "description": "Transfiere un turno creando uno nuevo asociado al servicio destino y actualizando el viejo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Transferir turno a otro servicio",
                "parameters": [
                    {
                        "description": "Datos de la transferencia",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/turns/call": {
            "post": {
                "description": "Actualiza el estado de un turno a llamado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Llamar un turno",
                "parameters": [
                    {
                        "description": "Datos para llamar turno",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.LlamadoTurnoPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.LlamadoTurnoPostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/turns/call/{turno}/{servicio}": {
            "get": {
                "description": "Verifica si un turno ha sido llamado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TurnosNTLIS"
                ],
                "summary": "Verificar llamado de turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de Turno",
                        "name": "turno",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de Servicio",
                        "name": "servicio",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.LlamadoTurnoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.AtencionEstadoRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "pointOfCare": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.PointOfCare"
                },
                "service": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.Service"
                },
                "state": {
                    "type": "integer",
                    "example": 3
                },
                "turn": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.Turn"
                },
                "user": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.User"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.AtencionEstadoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.SedesNTResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.SedeNT"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillasEstadoRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "integer",
                    "example": 1
                },
                "branch": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.Branch"
                },
                "point": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.Point"
                },
                "user": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.User"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillasEstadoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TaquillasEstadoRequest"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_TurnosNTLIS_domain.TransferResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}